requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.40
werkzeug>=3.1.3
wtforms>=3.2.1
flask-migrate>=4.1.0
python-dateutil>=2.9.0.post0
flask-wtf>=1.2.2
flask-dance>=7.1.0
oauthlib>=3.2.2
pyjwt>=2.10.1
anthropic
facebook-sdk
requests
numpy
pydub
moviepy
pillow
gTTS